# openapi-adminapi.yml
openapi: 3.1.0
info:
  title: MobileCMS Admin API
  description: Administrative operations for content management and theme configuration.
  version: v50
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
tags:
  - name: Admin
    description: Administrative operations
paths:
  /mobilecmsapi/v50/adminapi/theme:
    get:
      tags:
        - Admin
      operationId: getTheme
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
        - Admin
      operationId: postTheme
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonNode'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /mobilecmsapi/v50/adminapi/theme/banner:
    post:
      tags:
        - Admin
      operationId: banner
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                uploadfiles:
                  type: array
                  items:
                    type: string
                    format: binary
                banner:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /mobilecmsapi/v50/adminapi/index/{type}:
    get:
      tags:
        - Admin
      operationId: index_1
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
        - Admin
      operationId: rebuild_1
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /mobilecmsapi/v50/adminapi/content/{type}:
    get:
      tags:
        - Admin
      operationId: list_2
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
        - Admin
      operationId: create
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /mobilecmsapi/v50/adminapi/content/{type}/{id}:
    get:
      tags:
        - Admin
      operationId: get
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
        - Admin
      operationId: reset
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
        - Admin
      operationId: delete_1
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /mobilecmsapi/v50/adminapi/metadata/{type}:
    get:
      tags:
        - Admin
      operationId: metadata_1
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /mobilecmsapi/v50/adminapi/content:
    get:
      tags:
        - Admin
      operationId: types_1
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    JsonNode: {}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT